Move UnexpectedRowsExpectation to core directory
