+++ great_expectations/expectations/expectation.py
@@ -10,7 +10,6 @@ from collections import Counter
 from copy import deepcopy
 from inspect import isabstract
 from numbers import Number
-from string import Formatter
 from typing import (
     TYPE_CHECKING,
     Any,
@@ -1630,53 +1629,6 @@ representation."""  # noqa: E501
         success = bool(above_min and below_max)
 
         return {"success": success, "result": {"observed_value": metric_value}}
-
-
-class UnexpectedRowsExpectation(BatchExpectation):
-    """
-    UnexpectedRowsExpectations facilitate the execution of SQL or Spark-SQL queries as the core logic for an Expectation.
-
-    UnexpectedRowsExpectations must implement a `_validate(...)` method containing logic for determining whether data returned by the executed query is successfully validated.
-    One is written by default, but can be overridden.
-    A successful validation is one where the unexpected_rows_query returns no rows.
-
-    Args:
-        unexpected_rows_query (str): A SQL or Spark-SQL query to be executed for validation.
-    """  # noqa: E501
-
-    unexpected_rows_query: str
-
-    metric_dependencies: ClassVar[Tuple[str, ...]] = ("unexpected_rows_query.table",)
-    success_keys: ClassVar[Tuple[str, ...]] = ("unexpected_rows_query",)
-    domain_keys: ClassVar[Tuple[str, ...]] = (
-        "batch_id",
-        "row_condition",
-        "condition_parser",
-    )
-
-    @pydantic.validator("unexpected_rows_query")
-    def _validate_query(cls, query: str) -> str:
-        parsed_fields = [f[1] for f in Formatter().parse(query)]
-        if "batch" not in parsed_fields:
-            raise ValueError("Query must contain {batch} parameter.")  # noqa: TRY003
-
-        return query
-
-    @override
-    def _validate(
-        self,
-        metrics: dict,
-        runtime_configuration: dict | None = None,
-        execution_engine: ExecutionEngine | None = None,
-    ) -> Union[ExpectationValidationResult, dict]:
-        metric_value = metrics["unexpected_rows_query.table"]
-        return {
-            "success": len(metric_value) == 0,
-            "result": {
-                "observed_value": len(metric_value),
-                "details": {"unexpected_rows": metric_value},
-            },
-        }
 
 
 class QueryExpectation(BatchExpectation, ABC):
